Second big commit

+++ common/Messages/Types/SimpleTypes.h
@@ -11,12 +11,12 @@ namespace Messages {
     namespace SimpleTypes {
         typedef char int8;
         typedef unsigned char uint8;
-        typedef short int int16;
-        typedef unsigned short int uint16;
-        typedef long int int32;
-        typedef unsigned long int uint32;
-        typedef long long int64;
-        typedef unsigned long long uint64;
+        typedef int16_t int16;
+        typedef uint16_t uint16;
+        typedef int32_t int32;
+        typedef uint32_t uint32;
+        typedef int64_t int64;
+        typedef uint64_t uint64;
         typedef std::string string;
     }
 }
